+++ src/main/kotlin/blackjack/controller/GameController.kt
@@ -1,11 +1,11 @@
 package blackjack.controller
 
 import blackjack.domain.model.Bet
-import blackjack.domain.model.Dealer
-import blackjack.domain.model.Deck
 import blackjack.domain.model.Game
-import blackjack.domain.model.Hand
-import blackjack.domain.model.Player
+import blackjack.domain.model.card.Deck
+import blackjack.domain.model.card.Hand
+import blackjack.domain.model.participant.Dealer
+import blackjack.domain.model.participant.Player
 import blackjack.view.InputView
 import blackjack.view.OutputView
