+++ src/main/kotlin/blackjack/domain/model/Game.kt
@@ -1,5 +1,10 @@
 package blackjack.domain.model
 
+import blackjack.domain.model.card.Card
+import blackjack.domain.model.card.Deck
+import blackjack.domain.model.participant.Dealer
+import blackjack.domain.model.participant.Player
+
 class Game(
     private val deck: Deck,
     private val dealer: Dealer,
