+++ src/main/kotlin/blackjack/domain/model/participant/Participant.kt
@@ -1,4 +1,7 @@
-package blackjack.domain.model
+package blackjack.domain.model.participant
+
+import blackjack.domain.model.card.Card
+import blackjack.domain.model.card.Hand
 
 abstract class Participant(val name: String, cards: List<Card>) {
     protected val hand = Hand(cards)
